backend/requirements.txt
# Core API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
python-multipart>=0.0.9  # For file uploads

# AI Frameworks (Updated for 2025 compatibility)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0        # Updated for Chroma deprecation
langgraph>=0.1.0  # Critical update for cyclic graphs
vercel-ai==0.1.0

# Model Providers
langchain-google-genai>=1.0.3  # Gemini
langchain-anthropic>=0.1.11    # Claude (Legacy support)
langchain-groq>=0.1.3          # Groq (Llama 3)
langchain-openai>=0.1.0        # OpenAI / GitHub Models API
langchain-ollama>=0.1.0        # Ollama (Local models)

# Database & Storage
chromadb>=0.5.0                # Vector Store
supabase>=2.4.0                # Auth & User DB
redis>=5.0.4                   # Caching
neo4j>=5.15.0                  # Knowledge Graph (GraphRAG)

# Tools & Sandbox
e2b_code_interpreter>=0.0.10   # Python Sandbox (Required for Tutor)
langfuse>=2.30.0               # Observability

# ETL & Parsing (For Blackboard Zips)
aiofiles>=23.2.1
PyPDF2>=3.0.1
python-docx>=1.1.2
xmltodict>=0.13.0              # Easier than lxml for imsmanifest.xml
beautifulsoup4>=4.12.0         # For HTML parsing in .dat files
pandas>=2.2.0                  # For data manipulation

# Authentication & Security
PyJWT>=2.8.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
